Corrected Circle drawing algorithm
I was using this library on on a NodeMCU 1.0 with the vsCode PlatformIO development environment and I noticed that when I ran the basic example code, the circles were squashed (somewhat more like rounded rectangles than circles).  After a few hours of work comparing it to the Adafruit GFX library code and the wikipedia code for the midpoint circle algorithm, I determined that the pre-increments and pre-decrements used in all three circle drawing functions (drawCircle, drawCircleQuads, and fillCircle) should be post-increments and post-decrements.  This commit changes the two ++x's and one --y to x++ and y-- respectively in all three funtions.

With this change, this library now draws circles that match pixel by pixel with the calculations used in the Adafruit GFX library.  My testing has only covered smaller circles as is used in the example code but I believe it should scale up properly.