get rid of utf8ascii()
Make the drawString*() functions and getStringWidth() directly convert
UTF-8 on the fly if needed. This saves an extra malloc for the converted
string in most cases which then needs to be free()d and allows to count
drawn chars even for UTF-8 strings later.
Keep the utf8ascii() function to not break the API for derived classes.