add return values to drawString  & drawStringMaxWidth (#365)
* get rid of utf8ascii()

Make the drawString*() functions and getStringWidth() directly convert
UTF-8 on the fly if needed. This saves an extra malloc for the converted
string in most cases which then needs to be free()d and allows to count
drawn chars even for UTF-8 strings later.
Keep the utf8ascii() function to not break the API for derived classes.

* drawStringInternal: return number of chars drawn

Return the nuber of characters that was drawn. If this is less then the
string length, then the text was too long for the display. This allows
e.g. for custom word wrapping.

* drawString: return number of characters drawn

* drawStringMaxWidth: return chars written in first line

This allows do scroll easily through longer texts, by noting the number
of chars drawn in first line and then starting the text with this offset
in the next display cycle

* drawStringMaxWidth: fix UTF-8 width calculation

* add SSD1306ScrollVerticalDemo example

this shows how the return value of drawStringMaxWidth() can be used