Fix drawLogBuffer for scrolling
When logBuffer has logBufferMaxLines, the drawing needs to start shifted up by displayHeight % lineHeight so that the first line drops off a bit, and not the last. Now printing to display really just works.

(I moved the `length++` and removed the `else` because both sides of the condition started with it.)